# openapi.yaml
...
openapi: 3.1.0
info:
  title: SMIRK API
  version: "1.0.0"
  description: Route inventory generated from concrete Express route declarations in server.ts and src/team-routes.ts. Middleware app.use declarations are excluded from paths.
servers:
  - url: https://smirkcalls.com
    description: Production
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    WorkspaceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Workspace session or invite token
    ProvisioningBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Provisioning token
paths:
...
  "/api/sms/incoming":
    post:
      operationId: "POST_api_sms_incoming"
      tags:
        - "sms"
      summary: "POST /api/sms/incoming"
      x-source: "src/routes/sms-routes.ts:103"
      x-express-path: "/api/sms/incoming"
      security: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/sms/safety":
    get:
      operationId: "GET_api_sms_safety"
      tags:
        - "sms"
      summary: "GET /api/sms/safety"
      x-source: "src/routes/sms-routes.ts:42"
      x-express-path: "/api/sms/safety"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/sms/status":
    post:
      operationId: "POST_api_sms_status"
      tags:
        - "sms"
      summary: "POST /api/sms/status"
      x-source: "src/routes/sms-routes.ts:121"
      x-express-path: "/api/sms/status"
      security: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/sms/test":
    post:
      operationId: "POST_api_sms_test"
      tags:
        - "sms"
      summary: "POST /api/sms/test"
      x-source: "src/routes/sms-routes.ts:73"
      x-express-path: "/api/sms/test"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...